+++ tests/Components/GetSet/PostTest.php
@@ -21,8 +21,6 @@ class PostTest extends \Parable\Tests\Base
 
     public function testGetSetOnResource()
     {
-        $this->getSet->reset();
-
         $this->getSet->set("stuff", "here");
 
         $this->assertSame(